Add createdBy field, and allow to edit only for creator

+++ src/main/java/io/pogorzelski/nitro/carriers/service/RatingService.java
@@ -6,11 +6,13 @@ import io.pogorzelski.nitro.carriers.domain.Person;
 import io.pogorzelski.nitro.carriers.domain.Rating;
 import io.pogorzelski.nitro.carriers.repository.RatingRepository;
 import io.pogorzelski.nitro.carriers.repository.search.RatingSearchRepository;
+import io.pogorzelski.nitro.carriers.security.SecurityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -48,6 +50,9 @@ public class RatingService {
      */
     public Rating save(Rating rating) {
         log.debug("Request to save Rating : {}", rating);
+        if (rating.getId() != null){
+            checkPermissions(rating.getId());
+        }
         Country chargeCountry = rating.getChargeCountry();
         if (chargeCountry != null && chargeCountry.getId() != null){
             rating.setChargeCountry(entityManager.merge(chargeCountry));
@@ -101,7 +106,9 @@ public class RatingService {
      * @param id the id of the entity
      */
     public void delete(Long id) {
-        log.debug("Request to delete Rating : {}", id);        ratingRepository.deleteById(id);
+        log.debug("Request to delete Rating : {}", id);
+        checkPermissions(id);
+        ratingRepository.deleteById(id);
         ratingSearchRepository.deleteById(id);
     }
 
@@ -116,4 +123,11 @@ public class RatingService {
     public Page<Rating> search(String query, Pageable pageable) {
         log.debug("Request to search for a page of Ratings for query {}", query);
         return ratingSearchRepository.search(queryStringQuery(query), pageable);    }
+
+    public void checkPermissions(Long id) {
+        ratingRepository.findByCreatedByIsCurrentUser().stream()
+            .filter(rating -> rating.getId().equals(id))
+            .findFirst()
+            .orElseThrow(() -> new AccessDeniedException("Not allowed to modify this rating."));
+    }
 }
